Create gcp.py
This code works for creating and deleting instances, configuring the network and getting the external IP of the instance. Note that the default values are of my own project that I used to test the functionality. Some parts of the configuration have been commented and can serve as a skeleton code.